fix(tests): harden the shared FIFO read-count oracle against hangs (#819)
Port the cleanup hardening issue_796_domain_command_validation.rs's
independent copy grew during review -- nonblocking cleanup reader
descriptors, WriterOutcome/WriterMode, and the ReapedChild reap-on-ESRCH
guard -- into the shared support/fifo_snapshot.rs oracle every #808
snapshot control uses, and delete the now-duplicate implementation.
Adds a new negative control reproducing #813's development incident
(CLI exits before opening the FIFO path) to prove cleanup now
terminates instead of hanging.
