+++ rust/fslc/tests/support/fifo_snapshot.rs
@@ -1,22 +1,41 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Ryoichi Izumita
 
-//! Shared Unix CLI-level #808 read-count oracle.
+//! Shared Unix CLI-level #808/#796 read-count oracle.
 //!
 //! `TwoSnapshotFifo` binds a first writer to a first-created FIFO inode
 //! while atomically replacing the pathname with a second FIFO, so the CLI
 //! process cannot see source A concatenated with source B: only a second
 //! `open` of the pathname would reach B. Extracted from
 //! `issue_808_run_verify_snapshot.rs` (PR 1, #811) so
-//! `issue_808_testgen_scenarios_snapshot.rs` (PR 3, #808) can reuse the same
-//! oracle instead of re-deriving it. Pulled in per-file with
+//! `issue_808_testgen_scenarios_snapshot.rs` (PR 3, #808) and
+//! `issue_808_mutate_snapshot.rs` can reuse the same oracle instead of
+//! re-deriving it. Pulled in per-file with
 //! `#[path = "support/fifo_snapshot.rs"] mod fifo_snapshot;`, matching the
 //! existing `support/self_conformance_mapping.rs` convention -- not through
 //! `support/mod.rs`, which is unrelated corpus-walk tooling.
 //!
-//! A future #808 PR touching `mutate` may want the same oracle; extracting
-//! it here risks a merge conflict with that PR if it adds its own copy or
-//! its own extraction concurrently.
+//! #819 ported the cleanup hardening `issue_796_domain_command_validation.rs`
+//! grew during its own review into this shared copy and deleted that file's
+//! duplicate implementation, so all six FIFO-oracle tests now share one
+//! hardened oracle instead of an unhardened one drifting from a hardened one.
+//! Three properties motivate every piece of the hardening below:
+//!
+//! 1. `release_writer`'s cleanup path must never perform a *blocking* `open`
+//!    of the FIFO. Opening a FIFO for reading blocks until a writer attaches;
+//!    if the writer thread never reached its second `open` (it panicked, or
+//!    the CLI under test exited before opening the path at all -- see #813's
+//!    development, where a rejected CLI argument caused exactly this), a
+//!    blocking cleanup `open` hangs forever.
+//! 2. `Drop` funnels into that same cleanup path via `catch_unwind`, which
+//!    cannot interrupt a blocking syscall. A blocking `open` in `Drop` turns
+//!    an assertion failure mid-test into a silent hang instead of a reported
+//!    failure.
+//! 3. A timed-out child must still be reaped even if `kill` itself fails
+//!    (e.g. `ESRCH` because the child exited between the timeout poll and the
+//!    kill): `ReapedChild::terminate_and_reap_with` calls `wait`
+//!    unconditionally, and `Drop for ReapedChild` is the backstop if the
+//!    caller never reaps explicitly.
 
 #![cfg(unix)]
 
@@ -59,6 +78,28 @@ fn create_fifo(path: &Path) {
     assert!(status.success(), "mkfifo failed with {status}");
 }
 
+/// Outcome of the FIFO writer thread, reported by `release_writer` and by
+/// `Drop`'s best-effort cleanup so callers can tell an orderly finish apart
+/// from a caught writer panic.
+#[allow(dead_code)]
+#[derive(Debug, Eq, PartialEq)]
+pub enum WriterOutcome {
+    Finished,
+    Panicked,
+}
+
+/// Selects whether the writer thread completes normally or panics after
+/// writing source A but before the atomic rename to the replacement FIFO.
+/// `PanicBeforeRename` drives the cleanup-hardening control that proves
+/// `release_writer` terminates even when the writer never reaches its second
+/// `open`.
+#[allow(dead_code)]
+#[derive(Clone, Copy)]
+pub enum WriterMode {
+    Normal,
+    PanicBeforeRename,
+}
+
 /// The first writer remains attached to the first inode while the pathname is
 /// atomically replaced with the second FIFO. Thus source A cannot be
 /// concatenated with source B: only a second open of the pathname reaches B.
@@ -67,12 +108,19 @@ pub struct TwoSnapshotFifo {
     directory: PathBuf,
     pub fifo: PathBuf,
     second_opened: std::sync::mpsc::Receiver<()>,
+    writer_outcome: std::sync::mpsc::Receiver<WriterOutcome>,
+    phase: std::sync::Arc<std::sync::atomic::AtomicUsize>,
     writer: Option<std::thread::JoinHandle<()>>,
     second_writer_opened: bool,
 }
 
 #[allow(dead_code)]
 impl TwoSnapshotFifo {
+    const FIRST_WRITER_OPENING: usize = 1;
+    const REPLACED: usize = 2;
+    const SECOND_WRITER_OPENING: usize = 3;
+    const CLEANUP_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(5);
+
     /// `label` distinguishes concurrent fixture directories across test
     /// binaries. `source_a` and `source_b` should normally be two *distinct
     /// valid* FSL documents (not merely valid vs. invalid): a caller wants to
@@ -80,8 +128,22 @@ impl TwoSnapshotFifo {
     /// and got a different-but-still-valid B would otherwise look identical
     /// to a correct run under a status/`result != "error"` check alone.
     pub fn new(label: &str, source_a: &str, source_b: &str) -> Self {
+        Self::new_with_writer_mode(label, source_a, source_b, WriterMode::Normal)
+    }
+
+    /// As `new`, but drives the writer thread through `mode`. Used by the
+    /// cleanup-hardening control to force a writer panic before the FIFO path
+    /// is ever replaced, so `release_writer` must terminate without a second
+    /// `open` ever becoming reachable.
+    pub fn new_with_writer_mode(
+        label: &str,
+        source_a: &str,
+        source_b: &str,
+        mode: WriterMode,
+    ) -> Self {
         use std::io::Write;
-        use std::sync::mpsc;
+        use std::sync::atomic::Ordering;
+        use std::sync::{Arc, mpsc};
 
         let directory = fixture_directory(label);
         let fifo = directory.join("input.fsl");
@@ -92,34 +154,56 @@ impl TwoSnapshotFifo {
         let source_a = source_a.to_owned();
         let source_b = source_b.to_owned();
         let writer_fifo = fifo.clone();
-        let (second_opened, receiver) = mpsc::channel();
+        let phase = Arc::new(std::sync::atomic::AtomicUsize::new(0));
+        let writer_phase = Arc::clone(&phase);
+        let (second_opened, second_opened_receiver) = mpsc::channel();
+        let (writer_outcome, writer_outcome_receiver) = mpsc::channel();
         let writer = std::thread::spawn(move || {
-            let mut first = std::fs::OpenOptions::new()
-                .write(true)
-                .open(&writer_fifo)
-                .expect("open first FIFO for source A");
-            first
-                .write_all(source_a.as_bytes())
-                .expect("write source A");
-            std::fs::rename(&replacement, &writer_fifo).expect("replace FIFO path");
-            drop(first);
+            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
+                writer_phase.store(Self::FIRST_WRITER_OPENING, Ordering::SeqCst);
+                let mut first = std::fs::OpenOptions::new()
+                    .write(true)
+                    .open(&writer_fifo)
+                    .expect("open first FIFO for source A");
+                first
+                    .write_all(source_a.as_bytes())
+                    .expect("write source A");
+                if matches!(mode, WriterMode::PanicBeforeRename) {
+                    panic!("intentional FIFO writer panic before rename");
+                }
 
-            let mut second = std::fs::OpenOptions::new()
-                .write(true)
-                .open(&writer_fifo)
-                .expect("open replacement FIFO for source B");
-            second_opened
-                .send(())
-                .expect("test must retain second-open receiver");
-            second
-                .write_all(source_b.as_bytes())
-                .expect("write source B");
+                // `rename` replaces the directory entry, while `first` remains
+                // connected to the old FIFO inode until this explicit drop.
+                std::fs::rename(&replacement, &writer_fifo).expect("replace FIFO path");
+                writer_phase.store(Self::REPLACED, Ordering::SeqCst);
+                drop(first);
+
+                writer_phase.store(Self::SECOND_WRITER_OPENING, Ordering::SeqCst);
+                let mut second = std::fs::OpenOptions::new()
+                    .write(true)
+                    .open(&writer_fifo)
+                    .expect("open replacement FIFO for source B");
+                second_opened
+                    .send(())
+                    .expect("test must retain second-open receiver");
+                second
+                    .write_all(source_b.as_bytes())
+                    .expect("write source B");
+            }));
+            let outcome = if result.is_ok() {
+                WriterOutcome::Finished
+            } else {
+                WriterOutcome::Panicked
+            };
+            let _ = writer_outcome.send(outcome);
         });
 
         Self {
             directory,
             fifo,
-            second_opened: receiver,
+            second_opened: second_opened_receiver,
+            writer_outcome: writer_outcome_receiver,
+            phase,
             writer: Some(writer),
             second_writer_opened: false,
         }
@@ -143,35 +227,113 @@ impl TwoSnapshotFifo {
         }
     }
 
-    pub fn release_writer(&mut self) {
+    /// A nonblocking reader on the current FIFO path. `O_NONBLOCK` makes this
+    /// `open` return immediately (with `ENXIO`-free semantics on a FIFO that
+    /// already has a writer, or simply no data yet) instead of blocking until
+    /// a writer attaches -- the property a cleanup path must have, since a
+    /// writer may never attach at all (see the module doc's point 1).
+    fn open_nonblocking_control_fifo(&self) -> std::io::Result<std::fs::File> {
+        use std::os::unix::fs::OpenOptionsExt;
+
+        std::fs::OpenOptions::new()
+            .read(true)
+            .custom_flags(libc::O_NONBLOCK)
+            .open(&self.fifo)
+    }
+
+    fn drain_nonblocking_control_fifo(control: &mut std::fs::File) -> std::io::Result<()> {
         use std::io::Read;
 
+        let mut bytes = [0; 4096];
+        loop {
+            match control.read(&mut bytes) {
+                Ok(0) => return Ok(()),
+                Ok(_) => {}
+                Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => return Ok(()),
+                Err(error) => return Err(error),
+            }
+        }
+    }
+
+    /// Releases the writer thread without ever performing a blocking `open`.
+    /// Holds nonblocking reader descriptors on both the original and (once
+    /// reachable) the replacement FIFO path, draining them so the writer's
+    /// blocking `open`/`write_all` calls can complete, and bounds the wait by
+    /// `CLEANUP_TIMEOUT` instead of joining unconditionally. This cannot hang
+    /// even if the writer panics before ever reaching the rename: the
+    /// nonblocking descriptors do not depend on a second writer existing, and
+    /// the loop below only joins once `writer.is_finished()` is true.
+    pub fn release_writer(&mut self) -> WriterOutcome {
+        use std::sync::atomic::Ordering;
+        use std::sync::mpsc::TryRecvError;
+
         let Some(writer) = self.writer.take() else {
-            return;
+            return WriterOutcome::Finished;
         };
         if !self.second_writer_opened {
             self.second_writer_opened = self.second_opened.try_recv().is_ok();
         }
-        if !self.second_writer_opened {
-            let mut reader = std::fs::OpenOptions::new()
-                .read(true)
-                .open(&self.fifo)
-                .expect("open replacement FIFO during cleanup");
-            let mut source = String::new();
-            reader
-                .read_to_string(&mut source)
-                .expect("drain replacement FIFO during cleanup");
+
+        let mut controls = vec![
+            self.open_nonblocking_control_fifo()
+                .expect("open nonblocking FIFO cleanup control"),
+        ];
+        let mut replacement_control_opened = self.phase.load(Ordering::SeqCst) >= Self::REPLACED;
+        let deadline = std::time::Instant::now() + Self::CLEANUP_TIMEOUT;
+        let mut outcome = None;
+
+        loop {
+            if !replacement_control_opened && self.phase.load(Ordering::SeqCst) >= Self::REPLACED {
+                controls.push(
+                    self.open_nonblocking_control_fifo()
+                        .expect("open replacement FIFO cleanup control"),
+                );
+                replacement_control_opened = true;
+            }
+
+            for control in &mut controls {
+                Self::drain_nonblocking_control_fifo(control)
+                    .expect("drain nonblocking FIFO cleanup control");
+            }
+
+            if outcome.is_none() {
+                match self.writer_outcome.try_recv() {
+                    Ok(received) => outcome = Some(received),
+                    Err(TryRecvError::Empty) => {}
+                    Err(TryRecvError::Disconnected) => {
+                        panic!("FIFO writer exited without reporting an outcome");
+                    }
+                }
+            }
+
+            if writer.is_finished() {
+                let outcome = outcome.expect("finished FIFO writer must report an outcome");
+                writer.join().expect("join finished FIFO writer");
+                return outcome;
+            }
+
+            assert!(
+                std::time::Instant::now() < deadline,
+                "FIFO writer cleanup exceeded {:?}",
+                Self::CLEANUP_TIMEOUT
+            );
+            std::thread::sleep(std::time::Duration::from_millis(1));
         }
-        writer.join().expect("join FIFO writer");
     }
 }
 
 #[allow(dead_code)]
 impl Drop for TwoSnapshotFifo {
     fn drop(&mut self) {
+        // This runs during assertion unwinding as well, so the fixture cannot
+        // leave FIFOs (or a hung writer thread) behind merely because a check
+        // failed. `catch_unwind` only guards against `release_writer` itself
+        // panicking (e.g. its internal timeout assertion); it does not need
+        // to guard against a blocking syscall because `release_writer` no
+        // longer performs one.
         if self.writer.is_some() {
             let _ = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                self.release_writer();
+                let _ = self.release_writer();
             }));
         }
         let _ = std::fs::remove_file(&self.fifo);
@@ -179,39 +341,115 @@ impl Drop for TwoSnapshotFifo {
     }
 }
 
+/// Wraps a spawned child so a reap is never skipped, including when `kill`
+/// itself fails (e.g. `ESRCH` because the child already exited) and including
+/// on drop if the caller never reaps explicitly.
 #[allow(dead_code)]
-pub fn wait_for_output(child: &mut std::process::Child) -> std::process::Output {
-    use std::io::Read;
+pub struct ReapedChild {
+    pub child: std::process::Child,
+    reaped: bool,
+}
 
+#[allow(dead_code)]
+impl ReapedChild {
+    pub fn new(child: std::process::Child) -> Self {
+        Self {
+            child,
+            reaped: false,
+        }
+    }
+
+    pub fn try_wait(&mut self) -> std::io::Result<Option<std::process::ExitStatus>> {
+        let status = self.child.try_wait()?;
+        if status.is_some() {
+            self.reaped = true;
+        }
+        Ok(status)
+    }
+
+    pub fn wait(&mut self) -> std::io::Result<std::process::ExitStatus> {
+        let status = self.child.wait()?;
+        self.reaped = true;
+        Ok(status)
+    }
+
+    /// `wait` is deliberately unconditional: a process can exit between a
+    /// timeout poll and kill, making kill return `ESRCH` while still leaving
+    /// a child for this parent to reap. Returning the terminate error
+    /// separately (rather than `.expect()`-ing it) is exactly what lets the
+    /// reap proceed regardless of whether termination itself succeeded.
+    pub fn terminate_and_reap_with<F>(
+        &mut self,
+        terminate: F,
+    ) -> (
+        Option<std::io::Error>,
+        std::io::Result<std::process::ExitStatus>,
+    )
+    where
+        F: FnOnce(&mut std::process::Child) -> std::io::Result<()>,
+    {
+        let terminate_error = terminate(&mut self.child).err();
+        let status = self.wait();
+        (terminate_error, status)
+    }
+
+    fn capture_output(&mut self, status: std::process::ExitStatus) -> std::process::Output {
+        use std::io::Read;
+
+        let mut stdout = Vec::new();
+        let mut stderr = Vec::new();
+        self.child
+            .stdout
+            .take()
+            .expect("capture FIFO child stdout")
+            .read_to_end(&mut stdout)
+            .expect("read FIFO child stdout");
+        self.child
+            .stderr
+            .take()
+            .expect("capture FIFO child stderr")
+            .read_to_end(&mut stderr)
+            .expect("read FIFO child stderr");
+        std::process::Output {
+            status,
+            stdout,
+            stderr,
+        }
+    }
+}
+
+#[allow(dead_code)]
+impl Drop for ReapedChild {
+    fn drop(&mut self) {
+        if !self.reaped {
+            // Ignore errors during unwinding, but never let a failed kill
+            // skip the reap attempt.
+            let _ = self.child.kill();
+            let _ = self.wait();
+        }
+    }
+}
+
+#[allow(dead_code)]
+pub fn wait_for_output(child: &mut ReapedChild) -> std::process::Output {
     let deadline = std::time::Instant::now() + std::time::Duration::from_secs(5);
     let status = loop {
         if let Some(status) = child.try_wait().expect("poll FIFO child") {
             break status;
         }
         if std::time::Instant::now() >= deadline {
-            child.kill().expect("kill timed-out FIFO child");
-            let status = child.wait().expect("reap timed-out FIFO child");
-            panic!("fslc against FIFO timed out after 5s; status={status}");
+            let (kill_error, status) = child.terminate_and_reap_with(std::process::Child::kill);
+            let status = status.unwrap_or_else(|wait_error| {
+                panic!("reap timed-out FIFO child after kill result {kill_error:?}: {wait_error}")
+            });
+            let output = child.capture_output(status);
+            panic!(
+                "fslc against FIFO timed out after 5s; kill_error={kill_error:?}; status={status}; stdout={}; stderr={}",
+                String::from_utf8_lossy(&output.stdout),
+                String::from_utf8_lossy(&output.stderr)
+            );
         }
         std::thread::sleep(std::time::Duration::from_millis(10));
     };
-    let mut stdout = Vec::new();
-    let mut stderr = Vec::new();
-    child
-        .stdout
-        .take()
-        .expect("capture FIFO child stdout")
-        .read_to_end(&mut stdout)
-        .expect("read FIFO child stdout");
-    child
-        .stderr
-        .take()
-        .expect("capture FIFO child stderr")
-        .read_to_end(&mut stderr)
-        .expect("read FIFO child stderr");
-    std::process::Output {
-        status,
-        stdout,
-        stderr,
-    }
+    child.capture_output(status)
 }
